[#84] Run netmap service in neofs-node app
