+++ cmd/neofs-node/morph.go
@@ -1,12 +1,9 @@
 package main
 
 import (
-	v2netmap "github.com/nspcc-dev/neofs-api-go/v2/netmap"
-	crypto "github.com/nspcc-dev/neofs-crypto"
 	"github.com/nspcc-dev/neofs-node/pkg/morph/client"
 	"github.com/nspcc-dev/neofs-node/pkg/morph/client/netmap"
 	"github.com/nspcc-dev/neofs-node/pkg/morph/client/netmap/wrapper"
-	"github.com/pkg/errors"
 )
 
 func initMorphComponents(c *cfg) {
@@ -30,14 +27,4 @@ func initMorphComponents(c *cfg) {
 
 	c.cfgObject.netMapStorage = wrap
 	c.cfgNetmap.wrapper = wrap
-}
-
-func bootstrapNode(c *cfg) {
-	peerInfo := new(v2netmap.NodeInfo)
-	peerInfo.SetAddress(c.localAddr.String())
-	peerInfo.SetPublicKey(crypto.MarshalPublicKey(&c.key.PublicKey))
-	peerInfo.SetAttributes(c.cfgNodeInfo.attributes)
-
-	err := c.cfgNetmap.wrapper.AddPeer(peerInfo)
-	fatalOnErr(errors.Wrap(err, "bootstrap error"))
 }
